requirements.txt
streamlit>=1.28.0
requests>=2.25.0
matplotlib>=3.5.0
pandas>=1.3.0
python-dateutil>=2.8.0
